perf: replace pwrite ring sleep_for(1µs) spin with Highway futex
The pwrite ring buffer used std::this_thread::sleep_for(1µs) to poll
for the previous slot's published_seq. Replace with
hwy::BlockUntilDifferent + hwy::WakeAll for precise wakeup.

Changes:
- OffsetSlot::published_seq: atomic<size_t> → atomic<uint32_t>
  (seq values are small; uint32_t required by Highway futex API)
- Wait loop: sleep_for(1µs) → BlockUntilDifferent(cur, published_seq)
- Publish: store + WakeAll to notify waiting workers

+++ src/writerthread.h
@@ -19,7 +19,7 @@ static constexpr int OFFSET_RING_SIZE = 512;
 
 struct alignas(64) OffsetSlot {
     std::atomic<size_t> cumulative_offset{0};
-    std::atomic<size_t> published_seq{SIZE_MAX};
+    std::atomic<uint32_t> published_seq{UINT32_MAX};
 };
 
 class WriterThread{
